Finish map

+++ search/models.py
@@ -30,8 +30,8 @@ class UserSearch(models.Model):
         validators=[MinValueValidator(500), MaxValueValidator(500_000)],
     )
 
-    lat = models.DecimalField(max_digits=9, decimal_places=6)
-    lng = models.DecimalField(max_digits=9, decimal_places=6)
+    lat = models.FloatField()
+    lng = models.FloatField()
 
 
 class UserSearchStage(models.Model):
